# openapi.yaml
openapi: "3.0.2"
info:
  title: "apibox"
  version: "0.1"
components:
  schemas:
...
    ContributionRequest:
      type: "object"
      required:
        - "amount"
        - "description"
        - "userId"
      properties:
        amount:
          type: "integer"
        description:
          type: "string"
        userId:
          type: "string"
    ContributionResponse:
      type: "object"
      required:
        - "amount"
        - "createdAt"
        - "description"
        - "modifiedAt"
        - "publicId"
        - "userId"
      properties:
        amount:
          type: "integer"
        createdAt:
          type: "string"
          format: "date-time"
        description:
          type: "string"
        modifiedAt:
          type: "string"
          format: "date-time"
        publicId:
          type: "string"
        userId:
          type: "string"
...
    ErrorResponse:
      type: "object"
      required:
        - "code"
        - "message"
      properties:
        code:
          type: "string"
        message:
          type: "string"
        data:
          oneOf:
            - type: "object"
            - type: "array"
paths:
...
  /contributions:
    get:
      summary: "get contributions"
      tags:
        - "contributions"
      description: "Get the list of contributions"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/ContributionResponse"
    post:
      summary: "create contribution"
      tags:
        - "contributions"
      description: "Create a new contribution"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContributionRequest"
      responses:
        "201":
          description: "Created"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContributionResponse"
        "422":
          description: "Unprocessable Entity"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /contributions/{publicId}:
    get:
      summary: "get single contribution"
      tags:
        - "contributions"
      description: "Get a single contribution"
      parameters:
        - in: "path"
          name: "publicId"
          required: true
          schema:
            type: "string"
          description: "public id of the contribution to get"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContributionResponse"
        "404":
          description: "Not Found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...